Update Objects Design and Update Ball Behavior
Lessened the thickness of paddles so it does not look too thick and
reduced the radius of ball so it does not look too big and sped up the
ball velocity. Ball will reset position to it's initial state when each
of the player/cpu scores a goal. Also fix error where too much arguments
are being passed to the method that resets the ball's position.