Update Paddle Bounce Mechanic
Bounce angle will now depend on the distance between the center y coord
of the paddle and the center y coord of the ball. if it is farther to
the center, then incoming angle will be reflected to a larger angle. If
the ball hits the middle of the paddle then the angle will be 180 deg
or 0 deg (the ball will go straight horizontally). Also improved the
initial direction of the ball so that it always directs to the paddle
and not wall sometimes.